refactor(sparkline): 抽取了堆叠数据的方法
